+++ lib/src/services/desktop_timer_bridge.dart
@@ -1,0 +1,2 @@
+export 'desktop_timer_bridge_stub.dart'
+    if (dart.library.js_interop) 'desktop_timer_bridge_web.dart';