+++ lib/src/services/timer_completion_chime.dart
@@ -1,0 +1,2 @@
+export 'timer_completion_chime_stub.dart'
+    if (dart.library.js_interop) 'timer_completion_chime_web.dart';